requirements.txt
openai==1.6.1
pdfminer.six==20221105
PyQt6==6.4.2
pyqt6-tools==6.4.2.3.3
python-dotenv==1.0.0
